# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Employee Management System API",
    "description": "REST API backing the Employee Management System's web and mobile clients, covering three modules: employee & identity (employee/department records and JWT login with Admin/Manager/Employee roles), attendance (self-service daily check-in/check-out and history), and leave (leave requests and the Manager/Admin approval workflow).",
    "version": "v1"
  },
  "paths": {
    "/api/attendance/check-in": {
      "post": {
        "tags": [
          "Attendance"
        ],
        "summary": "Checks the authenticated employee in for today.",
        "description": "Any authenticated role. Always acts on the caller's own id (from the JWT) — there is no\r\nway to check someone else in. `DeviceType` records where the check-in came from\r\n(`Web` or `Mobile`). Only one check-in per employee per calendar day is allowed;\r\na second attempt returns 409.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckInDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckInDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CheckInDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/attendance/check-out": {
      "post": {
        "tags": [
          "Attendance"
        ],
        "summary": "Checks the authenticated employee out for today.",
        "description": "Any authenticated role. Takes no request body — it closes the caller's own currently-open\r\nattendance record for today (the row created by check-in), rather than creating a new one.\r\nReturns 400 if the caller hasn't checked in today, and 409 if they've already checked out.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/attendance": {
      "get": {
        "tags": [
          "Attendance"
        ],
        "summary": "Lists attendance records, optionally filtered by employee, department, and date range.",
        "description": "Role-scoped visibility: an `Employee` always sees only their own records — the\r\nemployeeId and departmentId filters are silently\r\nignored for that role rather than honoured, since the caller could otherwise pass anyone's\r\nid. `Manager`/`Admin` may filter freely across the whole company.\r\nstartDate/endDate are inclusive on both ends and\r\napply to every role.",
        "parameters": [
          {
            "name": "employeeId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "departmentId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AttendanceResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AttendanceResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "employeeId": {
            "type": "integer",
            "format": "int32"
          },
          "employeeName": {
            "type": "string",
            "nullable": true
          },
          "checkInTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "checkOutTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "workDate": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "deviceType": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CheckInDto": {
        "required": [
          "deviceType"
        ],
        "type": "object",
        "properties": {
          "deviceType": {
            "$ref": "#/components/schemas/DeviceType"
          }
        },
        "additionalProperties": false
      },
...
      "DeviceType": {
        "enum": [
          "Web",
          "Mobile"
        ],
        "type": "string",
        "description": "Origin of a check-in/check-out, since both the React web app and the mobile app write here."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": { }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT bearer token. Obtain one from POST /api/auth/login, then paste just the token here (Swagger UI adds the 'Bearer ' prefix for you).",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}